comparision from fragments integrated
