+++ src/pages/learn/useCases/index.js
@@ -1,23 +1,15 @@
 import React from 'react'
-import { Flex, Box } from 'serverless-design-system/src'
 
 import DefaultLayout from 'src/layouts/Default'
 
 import Prefooter from './Prefooter'
 import Hero from './Hero'
-
-import {
-  UseCasesList,
-  UseCasesSelector,
-} from '../../../fragments'
+import UseCaseScrollList from './UseCaseScrollList'
 
 const UseCases = () => (
   <DefaultLayout prefooter={Prefooter} transparentHeader>
     <Hero />
-    <Flex flexDirection={['column', 'column', 'row']}>
-      <UseCasesSelector />
-      <UseCasesList />
-    </Flex>
+    <UseCaseScrollList />
   </DefaultLayout>
 )
